fix(bloem): harden policy provisioning and repair contract fixtures
Keep explicit-organization account creation in its intended membership and
access group. Adopt only an active same-role seeded membership, and avoid
organization FK lock-upgrade deadlocks with FOR NO KEY UPDATE. Resolve the
login response's download capability from the newly authenticated account;
retain profileless account administration and selected-profile tenant/PIN
checks. Refuse API-key or sessionless logout, hide consumed invitations, and
restore established signup invite-code errors.

Use canonical, isolated fixtures and prerequisite overlays without changing
frozen scenarios, assertions or exclusions. Correct native engagement schema
nullability and client coverage metadata. Preserve audiobook attempt history
and accept bigint library filter parameters without relaxing access filters.

Add migration 20260919095615_bloem_policy_library_ids_bigint for the complete
policy-library array chain and rollback projections. Preserve authority,
policy data, revisions, array semantics and dependent trigger definitions,
deferral, enabled modes and comments. Guard Down against either int4 boundary
before narrowing any column. Quiesce writers and recycle pools for deployment;
image rollback should retain the schema. This is not universal wide-ID support.

Validation:
- Seven relevant package suites passed: 3,773 passing Go events; six existing
  handler skips. Three environment-gated checks subsequently passed against
  disposable dashboard/Redis resources, which were removed.
- Final migration/domain/tenant-wire/audiobook race selection: 31 passing
  events. Repeated migration/guard race tests: 27 passing events.
- Build, focused vet, formatting, native/client/route artifacts, migration
  validation and the 512-entry seam ledger passed. Scoped lint found zero
  changed/new-file issues; inherited full-package findings remain.
- The earlier complete Go run was NOT green: frozen route/scenario conflicts,
  local HDHomeRun broadcast failures and a handler failure. The handler suite
  now passes; no subsequent whole-repository green run is claimed. External
  Watch schemas and an inherited non-scalar setting fixture remain skipped.

The isolated production-backup restore and integrity drill passed separately;
it does not certify deployment or recovery of this new migration. No live
schema migration or deployment accompanies this commit.

AI disclosure: gpt-6-astra via pi coordinated implementation and review.
Earlier Codex CLI gpt-6-astra workers supplied bounded fixture/review patches;
subsequent fixes, migration and final review were coordinator-only. No Claude.
Review identified the organization lock cycle, invalid native nullability,
bigint policy overflow and trigger-comment loss; regressions cover the fixes.

+++ internal/audiobooks/media_store.go
@@ -290,8 +290,8 @@ func (s *ABSMediaStore) hydrateAudiobookRuntime(ctx context.Context, items []*mo
 			       SUM(COALESCE(duration, 0)) AS duration_seconds
 			FROM media_files
 			WHERE content_id = ANY($1) AND missing_since IS NULL
-			  AND ($2::int[] IS NULL OR media_folder_id = ANY($2))
-			  AND (COALESCE(cardinality($3::int[]), 0) = 0 OR NOT (media_folder_id = ANY($3)))
+			  AND ($2::bigint[] IS NULL OR media_folder_id = ANY($2))
+			  AND (COALESCE(cardinality($3::bigint[]), 0) = 0 OR NOT (media_folder_id = ANY($3)))
 			GROUP BY content_id, media_folder_id, COALESCE(presentation_group_key, '')
 		), chosen AS (
 			SELECT content_id, duration_seconds,
